feat(dashboard): extract dashboard content logic into separate components
Move data fetching and processing logic from page components to dedicated dashboardContent components for better separation of concerns and maintainability. Added Suspense with loading fallback for better user experience during data loading.

The changes include:
- Created dashboardContent.tsx for both user and creator dashboards
- Implemented loading states with Suspense
- Improved code organization by separating data logic from page layout
